+++ test_settings.py
@@ -20,8 +20,9 @@ def root(*args):
 
 AUTH_USER_MODEL = 'auth.User'
 CELERY_ALWAYS_EAGER = True
-COMPLETION_AGGREGATOR_BLOCK_TYPES = {'course', 'chapter', 'sequential'}
-COMPLETION_AGGREGATOR_ASYNC_AGGREGATION = True
+COMPLETION_AGGREGATOR_ASYNC_AGGREGATION = False
+COMPLETION_AGGREGATOR_BLOCK_TYPES = {'course', 'chapter', 'sequential', 'vertical'}
+COMPLETION_AGGREGATOR_TRACKING_EVENT_TYPES = COMPLETION_AGGREGATOR_BLOCK_TYPES
 COMPLETION_AGGREGATOR_AGGREGATION_LOCK = 'COMPLETION_AGGREGATOR_AGGREGATION_LOCK'
 COMPLETION_AGGREGATOR_CLEANUP_LOCK = 'COMPLETION_AGGREGATOR_CLEANUP_LOCK'
 COMPLETION_AGGREGATOR_AGGREGATION_LOCK_TIMEOUT_SECONDS = 1800
@@ -53,6 +54,7 @@ INSTALLED_APPS = (
     'oauth2_provider',
     'waffle',
     'test_utils.test_app',
+    'eventtracking.django.apps.EventTrackingConfig',
 )
 
 LOCALE_PATHS = [root('completion_aggregator', 'conf', 'locale')]
@@ -80,6 +82,9 @@ TEMPLATES = [
     },
 ]
 USE_TZ = True
+
+# Enables event tracking in the tests, see https://github.com/openedx/event-tracking
+EVENT_TRACKING_ENABLED = True
 
 # pylint: disable=unused-import,wrong-import-position
 from test_utils.test_app import celery  # isort:skip